move and rename components.py' module
move 'aea/configurations/components.py' into
'aea/components/ and rename it 'base.py'
