+++ aea/components/loader.py
@@ -29,7 +29,7 @@ from aea.configurations.base import (
     ComponentType,
     ConnectionConfig,
 )
-from aea.configurations.components import Component
+from aea.components.base import Component
 from aea.connections.base import Connection
 from aea.contracts.base import Contract
 from aea.exceptions import AEAPackageLoadingError
