+++ benchmark/framework/aea_test_wrapper.py
@@ -27,7 +27,7 @@ from benchmark.framework.fake_connection import FakeConnection
 from aea.aea import AEA
 from aea.aea_builder import AEABuilder
 from aea.configurations.base import SkillConfig
-from aea.configurations.components import Component
+from aea.components.base import Component
 from aea.crypto.fetchai import FetchAICrypto
 from aea.mail.base import Envelope
 from aea.protocols.base import Message
