+++ tests/test_aea_builder.py
@@ -27,7 +27,7 @@ import pytest
 
 from aea.aea_builder import AEABuilder
 from aea.configurations.base import ComponentType
-from aea.configurations.components import Component
+from aea.components.base import Component
 from aea.crypto.fetchai import FetchAICrypto
 from aea.exceptions import AEAException
